Check args prior to giving to clap.
If we encounter -exec, it is simply parsed to be --exec, this ensures
that exec works in this fashion, but sadly removes short arg support for
all those xec files.

Rename function, shorten names of types.

Also, do not bother running the map over if no entries are
specifically exec.