doc: Rewrite manpage with scdoc
I'm not entirely sure if this is worth it.

On the one hand, I think the source is clearer, easier to read, and
easier to modify.

However, scdoc has a few limitations and odd edge cases, that makes it a
little awkward to work with.

Another alternative is we could use ASCIIDOC with pandoc or asciidoctor
to convert it.